Extract folder tree widgets
